Przeniosłem obiekty z config do osobnych plików: konsultant.py, mailsender.py, klient.py Dodałem kilka komentarzy dla czytelności kodu.
