+++ klient.py
@@ -1,0 +1,34 @@
+'''Moduł zawierający klasę i obiekt Klient'''
+
+class Klient():
+    '''Klasa Klient dla obiektu klient,
+    przechowuje jego wszystkie dane.'''
+    def __init__(self):
+        pass
+
+    def stworz_klienta(self, imnaz, tel, datasprz, datawys,
+                       cena_dl, mail, branza, pytania, dodatkowe,
+                       rej_var, kor_var, dost_var,
+                       rej, kor, dost):
+        self.imnaz = imnaz
+        self.tel = tel
+        self.datasprz = datasprz
+        self.datawys = datawys
+        self.cena_dl = cena_dl
+        self.mail = mail
+        self.branza = branza
+        self.pytania = pytania
+        self.dodatkowe = dodatkowe
+        self.rej_var = rej_var
+        self.kor_var = kor_var
+        self.dost_var = dost_var
+        if not self.rej_var:
+            self.rej = rej
+        if not self.kor_var:
+            self.kor = kor
+        if not self.dost_var:
+            self.dost = dost
+
+
+
+klient = Klient()
